orders/requirements.txt
django
django-rest-passwordreset
psycopg2-binary
djangorestframework
pyyaml